fix: clippy -D warnings clean across 5 crates
- manual_div_ceil → .div_ceil() (messages.rs, evidence_store.rs)
- collapsible_if → let-chain (adapter, disk_store, epoch, timeout, proposal)
- needless_borrows_for_generic_args → remove & (disk_store, evidence_store)
- unnecessary_cast u8→u8 (proof.rs)
- for_kv_map → .keys() (sharded.rs, tree.rs)
- manual_contains → .contains() (epoch.rs)
- type_complexity → #[allow] (sharded.rs)
- dead_code: validator_set_for_view → #[cfg(test)] (adapter.rs)
- doc_lazy_continuation (chainspec.rs)

+++ crates/n42-jmt/src/proof.rs
@@ -110,7 +110,7 @@ pub fn build_proof<S: crate::TreeStore>(
 ) -> eyre::Result<Option<JmtProof>> {
     let start = Instant::now();
 
-    let shard_idx = (key_hash.0[0] >> 4) as u8;
+    let shard_idx = key_hash.0[0] >> 4;
 
     // Generate shard-level proof. None if shard is empty.
     let Some(shard_proof) = sharded_jmt.get_proof(key_hash)? else {
